Modified all Window event handlers that took an EventArgs& parameter to take an WindowEventArgs& parameter. Added documentation to Window event handlers. Modified firing of some events, added some others. Renamed Window::setRestoreCaptue method to setRestoreCapture. Added missing 'requestRedraw' call in ProgressBar::onProgressChanged. Fixed bug which rendered disabled state useless for all Windows. Fixed bugs in FrameWindow related to various options. Added code to TLFrameWindow to hide title bar and close button when they are disabled. Applied patches from Spoke to resolve some compiler warnings. Other minor changes.

+++ include/CEGUIImageset.h
@@ -124,7 +124,7 @@ public:
 	\return
 		uint value equal to the number of Image objects defined for the Imageset
 	*/
-	uint		getImageCount(void) const					{return d_images.size();}
+	uint      getImageCount(void) const               {return (uint)d_images.size();}
 
  
 	/*!
